Roll back split of PAF-lib modules
