+++ paf-mvp-frontend/rollup.config.js
@@ -34,7 +34,7 @@ const generateSourceMap = IS_DEV;
 // https://rollupjs.org/guide/en/#configuration-files
 export default [
   defineConfig({
-    input: relative('src/lib/paf-lib.ts'),
+    input: relative('src/lib/one-key.ts'),
     output: {
       file: getDestFolder(`/paf-lib.js`),
       format: 'umd',
