+++ paf-mvp-frontend/src/global.d.ts
@@ -1,5 +1,5 @@
 import { IProcessingQueue } from './utils/queue';
-import { IOneKeyLib } from '@frontend/lib/i-one-key-lib';
+import { IOneKeyLib } from '@frontend/lib/paf-lib';
 
 export type Window = WindowProxy &
   typeof globalThis & {
