--- paf-mvp-frontend/src/lib/rtb.ts
@@ -1,36 +1,0 @@
-import { IdsAndPreferences, Seed, TransactionId } from '@core/model';
-
-export type PrebidTransactionId = string;
-export type DivId = string;
-export type ContentId = string;
-
-export interface AuditHandler {
-  // Binds the element to the audit handler.
-  bind(element: HTMLElement): void;
-}
-
-export interface TransmissionRegistryContext {
-  /** Transaction Id generated by Prebidjs. */
-  prebidTransactionId: PrebidTransactionId;
-  /** Transaction Id generated for the OneKey Audit Log and used for signing the Seed. */
-  pafTransactionId: TransactionId;
-  /**
-   * The Id of the tag (<div id="something">) that contains the
-   * addressable content or the Google Publisher Tag adUnitCode.
-   */
-  divIdOrAdUnitCode: string;
-  contentId: ContentId;
-  auditHandler?: AuditHandler;
-}
-
-/**
- * Aggregate Seed and Preferences for storage.
- *
- * We must record the Ids and Preferences
- * in case they change in the cookies after the
- * generation of the Seed.
- */
-export interface SeedEntry {
-  seed: Seed;
-  idsAndPreferences: IdsAndPreferences;
-}
