+++ paf-mvp-frontend/tests/test-cases/onekey-lib.test.ts
@@ -11,8 +11,7 @@ import {
 import fetch from 'jest-fetch-mock';
 import { isBrowserKnownToSupport3PC } from '@core/user-agent';
 import { MockedFunction } from 'ts-jest';
-import { IdsAndPreferencesResult } from '@frontend/lib/i-one-key-lib';
-import { OneKeyLib } from '@frontend/lib/one-key-lib';
+import { IdsAndPreferencesResult, OneKeyLib } from '@frontend/lib/paf-lib';
 
 jest.mock('@core/user-agent', () => ({ isBrowserKnownToSupport3PC: jest.fn() }));
 jest.mock('ua-parser-js', () => () => ({ getBrowser: () => 'JEST-DOM' }));
